+++ test/integration/test_package.py
@@ -2,9 +2,6 @@ import os.path
 
 from . import *
 
-is_mingw = (env.host_platform.family == 'windows' and
-            env.builder('c++').flavor == 'cc')
-
 
 class TestPackage(IntegrationTest):
     def __init__(self, *args, **kwargs):
@@ -13,9 +10,7 @@ class TestPackage(IntegrationTest):
 
     def test_build(self):
         self.build()
-        # XXX: This fails on MinGW (not sure why)...
-        if not is_mingw:
-            self.assertOutput([executable('program')], '')
+        self.assertOutput([executable('program')], '')
 
 
 class TestSystemPackage(IntegrationTest):
@@ -25,12 +20,10 @@ class TestSystemPackage(IntegrationTest):
 
     def test_build(self):
         self.build()
-        # XXX: This fails on MinGW (not sure why)...
-        if not is_mingw:
-            self.assertOutput([executable('program')], '')
+        self.assertOutput([executable('program')], '')
 
 
-@skip_if(is_mingw, 'xfail on mingw')
+@skip_if('boost' not in test_features, 'skipping boost tests')
 class TestBoostPackage(IntegrationTest):
     def __init__(self, *args, **kwargs):
         super().__init__('boost', *args, **kwargs)