+++ doc/about/changes.md
@@ -6,6 +6,7 @@ in progress
 
 ### Bug fixes
 - Error reporting now shows the proper stack frames on Windows
+- Fix detection of `clang-cl` compiler
 
 ---